# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Permify is an open-source authorization service for creating and maintaining fine-grained authorizations accross your individual applications and services.\nPermify converts authorization data as relational tuples into a database you point at. We called that database a Write Database (WriteDB) and it behaves as a centralized data source for your authorization system. You can model of your authorization with Permify's DSL - Permify Schema - and perform access checks with a single API call anywhere on your stack. Access decisions made according to stored relational tuples.",
        "title": "Permify API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/Permify/permify/issues",
            "email": "hello@permify.co"
        },
        "license": {
            "name": "GNU Affero General Public License v3.0"
        },
        "version": "v0.0.0-alpha3"
    },
    "paths": {
...
        "/schemas/lookup": {
            "post": {
                "description": "lookup your authorization model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schema"
                ],
                "summary": "Schema",
                "operationId": "schemas.lookup",
                "parameters": [
                    {
                        "description": "lookup your authorization model",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.LookupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.LookupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/schemas/read": {
            "post": {
                "description": "read your authorization model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schema"
                ],
                "summary": "Schema",
                "operationId": "schemas.read",
                "parameters": [
                    {
                        "description": "read your authorization model",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.ReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.ReadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/schemas/write": {
            "post": {
                "description": "write your authorization model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schema"
                ],
                "summary": "Schema",
                "operationId": "schemas.write",
                "parameters": [
                    {
                        "description": "schema file (expected extension .perm)",
                        "name": "schema",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.WriteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.HTTPErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {}
            }
        },
...
        "schema.Action": {
            "type": "object",
            "properties": {
                "child": {},
                "name": {
                    "type": "string"
                }
            }
        },
        "schema.Entity": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Action"
                    }
                },
                "name": {
                    "type": "string"
                },
                "relations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Relation"
                    }
                }
            }
        },
        "schema.LookupRequest": {
            "type": "object",
            "properties": {
                "entity_type": {
                    "type": "string"
                },
                "relation_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "schema_version": {
                    "type": "string"
                }
            }
        },
        "schema.LookupResponse": {
            "type": "object",
            "properties": {
                "action_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "schema.ReadRequest": {
            "type": "object",
            "properties": {
                "schema_version": {
                    "type": "string"
                }
            }
        },
        "schema.ReadResponse": {
            "type": "object",
            "properties": {
                "entities": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/schema.Entity"
                    }
                }
            }
        },
        "schema.Relation": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "schema.WriteResponse": {
            "type": "object",
            "properties": {
                "version": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "api-key-auth",
            "in": "header"
        }
    }
}